break agentic app manager demo into separate component
